create consensus_message file
